+++ HISTORY.rst
@@ -9,11 +9,18 @@ New:
 - Add session timeout configuration to Security panel
   [lucid-0]
 
+- Add audio/video twitter cards
+  [vangheem, lucid-0]
+
+
 Fixes:
 
 - Handle error in `@@content-body` when there is no IFeedItem adapter
   for the current context
   [vangheem]
+
+- Fix twitter cards
+  [vangheem, lucid-0]
 
 
 2.3.6 (2018-12-20)